samples/: Use #get_open_file_url instead of #get_open_file_name
The latter is temporarily "removed" (That is, ignored by bindgen)
because it uses a `QString *` argument in its original C++ version.

Current versions of bindgen don't know how to pass in a pointer of
a possibly wrapped (Which is the case here) type.  If it detects a
method doing this, the method is ignored.