Always check/download configured Qt versions
Previously, existence of archive file in `download_cache/` was implicitly
taken to mean the file is valid and complete, and a download/resume wasn't
attempted.

This, if the file wasn't complete, made tar report unclear errors.

Now downloads are always attempted using curl option -C, which will cause
archive files to always be re-checked and downloads will potentially be
resumed. If the files are complete, then the checks will exit quickly.

This is a test to see if this will be more user friendly.