Qt::Application resets numeric locale to "C"
This is necessary (and adviced by Qt) because Qt sets the locale to
the system default when creating an application object. However,
Crystal functions like `to_f` depend on the locale being set to "C" by
default.

Because the locale is Unix-specific the reset is done by adding an
additional C helper function, which is called from
`Qt::Application#initialize`.